# core/core-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cloud Between API
  description: |
    구름 사이 (Cloud Between) 서비스 API 명세

    ## 서비스 개요
    "구름 사이"는 하늘과 구름을 테마로 한 심리 진단 서비스입니다.
    사용자는 12단계의 심리 테스트를 통해 자신의 구름 페르소나(Cloud Type)를 발견하고,
    다른 구름 유형과의 궁합(Chemistry)을 확인할 수 있습니다.

    ## 6가지 구름 유형 (Cloud Types)
    | Type Key | 이름 | 설명 |
    |----------|------|------|
    | `sunlit` | 햇살 구름 | 밝고 따뜻한 에너지, 주변을 환하게 만드는 존재 |
    | `mist` | 안개 구름 | 신비롭고 조용한 분위기, 깊은 내면의 세계 |
    | `storm` | 폭풍 구름 | 강렬한 감정과 에너지, 변화를 이끄는 힘 |
    | `dawn` | 새벽 구름 | 희망적이고 새로운 시작, 잔잔한 영감 |
    | `shade` | 그늘 구름 | 차분하고 보호적인 성향, 안정감을 주는 존재 |
    | `wild` | 야생 구름 | 자유롭고 예측 불가능, 창의적인 에너지 |

    ## 주요 플로우
    1. **심리 진단** (`/diagnosis`): 12단계 질문 → 답변 제출 → 구름 유형 결과
    2. **페르소나 조회** (`/personas`): 각 구름 유형의 상세 프로필 확인
    3. **궁합 조회** (`/chemistries`): 두 구름 유형 간의 관계성 확인

    ## 아키텍처 원칙
    - 도메인 간 격벽(Wall)을 위해 ORM 수준의 연관관계(Join)를 사용하지 않습니다.
    - 각 도메인은 독립적인 개념(Concept)으로 관리됩니다.
    - 비회원도 진단 테스트를 수행할 수 있으며, 회원은 결과가 저장됩니다.
  version: 1.0.0
servers:
  - url: https://cloud-between.duckdns.org/api/v1
    description: 운영 서버
  - url: http://localhost:8081/api/v1
    description: 로컬 서버
...
paths:
  /users/signup:
    post:
      tags: [User]
      summary: 회원가입
      description: 이메일과 비밀번호로 신규 계정을 생성합니다.
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: 가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /users/login:
    post:
      tags: [User]
      summary: 이메일/비밀번호 로그인
      description: 등록된 이메일과 비밀번호로 로그인합니다.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /users/login/social:
    post:
      tags: [User]
      summary: 소셜 로그인 (없으면 자동 회원가입)
      description: |
        소셜 로그인 플로우:
        1. social_id + provider로 기존 유저 조회
        2. 있으면 → 로그인 처리 (last_login_at 갱신)
        3. 없으면 → 자동 회원가입 후 유저 반환

        클라이언트는 가입/로그인 구분 없이 이 엔드포인트 하나만 호출하면 됩니다.
      operationId: socialLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialLoginRequest'
      responses:
        '200':
          description: 로그인 성공 (기존 유저)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '201':
          description: 자동 회원가입 후 로그인 성공 (신규 유저)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 지원하지 않는 provider 또는 잘못된 요청
  /diagnosis/questions:
    get:
      tags: [Diagnosis]
      summary: 심리 테스트 질문 목록 조회 (단계별)
      description: |
        12단계로 구성된 심리 테스트 질문을 단계(Step) 단위로 반환합니다.

        **응답 구조**: Step[] → 각 Step 안에 Question[] → 각 Question 안에 Option[]

        **프론트엔드 사용 방식**:
        - 각 Step에 포함된 questions 배열에서 랜덤으로 1개의 질문을 선택하여 사용자에게 표시
        - 총 12단계이므로, 사용자는 12개의 질문에 답변하게 됩니다
        - 각 질문의 선택지(options)에는 cloudType이 포함되어 있어 답변 집계에 사용됩니다

        **인증**: 불필요 (비회원도 조회 가능)
      operationId: getQuestions
      parameters:
        - name: locale
          in: query
          description: |
            다국어 코드. 해당 언어로 작성된 질문/선택지를 반환합니다.
            현재 지원: `ko` (한국어), `en` (영어)
          schema:
            type: string
            default: ko
            enum: [ko, en]
      responses:
        '200':
          description: 단계별 질문 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StepResponse'
              example:
                - id: 1
                  title: "감정 표현"
                  emoji: "💭"
                  questions:
                    - id: 1
                      text: "기분이 안 좋을 때 나는..."
                      options:
                        - text: "밝은 척 웃으며 넘긴다"
                          cloudType: "sunlit"
                        - text: "혼자 조용히 삭인다"
                          cloudType: "mist"
                        - text: "화를 바로 표현한다"
                          cloudType: "storm"
                        - text: "일기에 쓰거나 정리한다"
                          cloudType: "dawn"
                    - id: 2
                      text: "감정을 표현할 때 나는..."
                      options:
                        - text: "표정으로 다 드러난다"
                          cloudType: "sunlit"
                        - text: "글이나 예술로 표현한다"
                          cloudType: "mist"
  /diagnosis/analyze:
    post:
      tags: [Diagnosis]
      summary: 답변 기반 구름 유형 진단
      description: |
        사용자의 답변(12개)을 분석하여 가장 많이 선택된 구름 유형(cloudType)을 결과로 반환합니다.

        **진단 로직**:
        1. 각 답변의 cloudType을 집계합니다
        2. 가장 많이 선택된 cloudType이 최종 결과가 됩니다
        3. 결과의 personaType 필드로 반환됩니다

        **결과 저장**:
        - Authorization 헤더가 있는 경우(로그인 유저): 진단 결과가 DB에 저장됩니다
        - Authorization 헤더가 없는 경우(비회원): 결과만 반환하고 저장하지 않습니다

        **인증**: 선택적 (비회원도 진단 가능, 회원은 결과 저장)
      operationId: analyzeDiagnosis
      parameters:
        - name: Authorization
          in: header
          required: false
          description: |
            로그인 유저의 인증 토큰 (Bearer).
            있으면 진단 결과를 서버에 저장합니다. 없으면 결과만 반환합니다.
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIs..."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiagnosisRequest'
            example:
              answers:
                - questionId: 1
                  cloudType: "sunlit"
                - questionId: 5
                  cloudType: "mist"
                - questionId: 9
                  cloudType: "sunlit"
                - questionId: 13
                  cloudType: "storm"
                - questionId: 17
                  cloudType: "sunlit"
                - questionId: 21
                  cloudType: "dawn"
                - questionId: 25
                  cloudType: "sunlit"
                - questionId: 29
                  cloudType: "shade"
                - questionId: 33
                  cloudType: "wild"
                - questionId: 37
                  cloudType: "sunlit"
                - questionId: 41
                  cloudType: "mist"
                - questionId: 45
                  cloudType: "sunlit"
      responses:
        '200':
          description: |
            진단 결과 반환. personaType 필드에 가장 많이 선택된 구름 유형이 담깁니다.
            이 값으로 `/personas/profiles/{typeKey}` API를 호출하면 상세 프로필을 확인할 수 있습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosisResult'
              example:
                personaType: "sunlit"
        '400':
          description: 잘못된 요청 (answers 누락 또는 형식 오류)
...
  /chemistries/match:
    get:
      tags: [Chemistry]
      summary: 두 구름 유형의 궁합 조회
      description: |
        특정 두 구름 유형의 궁합 정보를 조회합니다.
        type1과 type2의 순서는 무관합니다 (sunlit+mist = mist+sunlit).

        **사용 사례**:
        - 진단 결과 후 친구와의 궁합 확인
        - 두 유형을 선택하여 궁합 상세 보기
      operationId: getChemistry
      parameters:
        - name: type1
          in: query
          required: true
          description: 첫 번째 구름 유형
          schema:
            type: string
            enum: [sunlit, mist, storm, dawn, shade, wild]
          example: sunlit
        - name: type2
          in: query
          required: true
          description: 두 번째 구름 유형
          schema:
            type: string
            enum: [sunlit, mist, storm, dawn, shade, wild]
          example: mist
      responses:
        '200':
          description: 궁합 결과 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChemistryResponse'
        '400':
          description: type1 또는 type2 누락
        '404':
          description: 해당 조합의 궁합 데이터가 존재하지 않음
...
  /premium-cards/{category}:
    get:
      tags: [PremiumCard]
      summary: 특정 카테고리의 프리미엄 카드 템플릿 조회
      description: |
        특정 카테고리의 프리미엄 카드 템플릿을 locale 기준으로 필터링하여 반환합니다.
        locale이 없는 데이터(themes 등)는 항상 포함됩니다.
      operationId: getPremiumCardsByCategory
      parameters:
        - name: category
          in: path
          required: true
          description: 템플릿 카테고리
          schema:
            type: string
            enum: [opening_patterns, softening_lines, strength_lines, growth_lines, closing_lines, share_card_phrases, share_card_captions, friction_patterns, misreading_patterns, themes]
          example: opening_patterns
        - name: locale
          in: query
          description: |
            다국어 코드. 해당 언어의 템플릿만 반환합니다.
            현재 지원: `ko` (한국어), `en` (영어)
          schema:
            type: string
            default: ko
            enum: [ko, en]
      responses:
        '200':
          description: 해당 카테고리의 템플릿 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PremiumCardResponse'
        '404':
          description: 해당 카테고리의 데이터가 존재하지 않음
  /translations:
    get:
      tags: [Translation]
      summary: 전체 번역 데이터 조회 (locale 기준)
      description: |
        특정 locale의 모든 번역 문자열을 namespace별로 그룹핑하여 반환합니다.

        **응답 구조**: `{ namespace: { keyPath: value } }`

        **사용 사례**:
        - 프론트엔드 초기 로딩 시 전체 번역 데이터 일괄 조회
        - 언어 전환 시 새 locale 데이터 로드
      operationId: getAllTranslations
      parameters:
        - name: locale
          in: query
          description: |
            다국어 코드. 해당 언어의 번역 문자열을 반환합니다.
            현재 지원: `ko` (한국어), `en` (영어)
          schema:
            type: string
            default: ko
            enum: [ko, en]
      responses:
        '200':
          description: namespace별로 그룹핑된 번역 데이터
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationAllResponse'
              example:
                home:
                  pageTitle: "Cloud Between Us — 우리 사이의 하늘"
                  hero.headline: "사랑은, 날씨처럼 변해."
                  hero.cta: "우리 하늘 보기"
                test:
                  pageTitle: "우리 사이의 하늘 — Cloud Between Us"
                  back: "이전으로"
                  next: "다음으로"
                auth:
                  login: "로그인"
                  signup: "회원가입"
  /translations/{namespace}:
    get:
      tags: [Translation]
      summary: 특정 namespace 번역 데이터 조회
      description: |
        특정 locale + namespace의 번역 문자열을 `{ keyPath: value }` 형태로 반환합니다.

        **사용 사례**:
        - 특정 페이지 진입 시 해당 페이지의 번역만 조회
        - Backoffice에서 namespace 단위 편집 화면
      operationId: getTranslationsByNamespace
      parameters:
        - name: namespace
          in: path
          required: true
          description: 번역 그룹 (페이지/섹션)
          schema:
            type: string
            enum: [home, test, result, auth, footer, lang]
          example: home
        - name: locale
          in: query
          description: |
            다국어 코드. 해당 언어의 번역 문자열을 반환합니다.
            현재 지원: `ko` (한국어), `en` (영어)
          schema:
            type: string
            default: ko
            enum: [ko, en]
      responses:
        '200':
          description: keyPath-value 형태의 번역 데이터
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationNamespaceResponse'
              example:
                pageTitle: "Cloud Between Us — 우리 사이의 하늘"
                hero.headline: "사랑은, 날씨처럼 변해."
                hero.headline2: "지금 너희 사이엔 어떤 공기가 흐르고 있어?"
                hero.subtext: "2분이면 충분해.\n둘 사이의 구름을 만나봐."
                hero.cta: "우리 하늘 보기"
                hero.noSignup: "가입 없이 바로."
                concept.title: "두 사람이 마주하기 전,\n먼저 공기가 바뀌어."
                clouds.title: "사랑이 흐르는 방식은 여섯 가지."
                clouds.names: "햇살,안개,천둥,여명,산들바람,그림자"
        '404':
          description: 해당 namespace의 번역 데이터가 존재하지 않음
  /personas/profiles:
    get:
      tags: [Persona]
      summary: 전체 페르소나 프로필 목록 조회
      description: |
        6가지 구름 유형의 상세 프로필 정보를 모두 반환합니다.

        각 프로필에는 다음 정보가 포함됩니다:
        - **typeKey**: 구름 유형 식별자 (sunlit, mist, storm, dawn, shade, wild)
        - **name**: 구름 유형의 한글 이름 (예: "햇살 구름")
        - **emoji**: 대표 이모지 (예: "☀️")
        - **subtitle**: 한 줄 설명
        - **keywords**: 성격 키워드 배열 (예: ["밝은", "따뜻한", "사교적인"])
        - **lore**: 구름 유형에 대한 서사/설명
        - **strengths**: 강점 목록
        - **shadows**: 그림자(약점) 목록

        **사용 사례**:
        - 전체 구름 유형 소개 페이지
        - 진단 결과 후 모든 유형 비교 화면
      operationId: getPersonaProfiles
      parameters:
        - name: locale
          in: query
          description: |
            다국어 코드. 해당 언어의 프로필 컨텐츠를 반환합니다.
            현재 지원: `ko` (한국어), `en` (영어)
          schema:
            type: string
            default: ko
            enum: [ko, en]
      responses:
        '200':
          description: 6가지 구름 유형의 프로필 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonaProfileResponse'
  /personas/profiles/{typeKey}:
    get:
      tags: [Persona]
      summary: 특정 구름 유형 프로필 조회
      description: |
        특정 구름 유형의 상세 프로필 정보를 반환합니다.
        진단 결과로 받은 personaType 값을 typeKey로 사용하여 상세 정보를 조회할 수 있습니다.

        **사용 사례**:
        - 진단 결과 화면에서 해당 유형의 상세 프로필 표시
        - 특정 구름 유형 상세 페이지
      operationId: getPersonaProfile
      parameters:
        - name: typeKey
          in: path
          required: true
          description: 구름 유형 식별자
          schema:
            type: string
            enum: [sunlit, mist, storm, dawn, shade, wild]
          example: sunlit
        - name: locale
          in: query
          description: |
            다국어 코드. 해당 언어의 프로필 컨텐츠를 반환합니다.
            현재 지원: `ko` (한국어), `en` (영어)
          schema:
            type: string
            default: ko
            enum: [ko, en]
      responses:
        '200':
          description: 해당 구름 유형의 프로필 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaProfileResponse'
        '404':
          description: 해당 typeKey의 페르소나를 찾을 수 없음
components:
  schemas:
    CloudType:
      type: string
      enum: [sunlit, mist, storm, dawn, shade, wild]
      description: |
        구름 유형 식별자.
        - `sunlit`: 햇살 구름 - 밝고 따뜻한 에너지
        - `mist`: 안개 구름 - 신비롭고 조용한 분위기
        - `storm`: 폭풍 구름 - 강렬한 감정과 에너지
        - `dawn`: 새벽 구름 - 희망적이고 새로운 시작
        - `shade`: 그늘 구름 - 차분하고 보호적인 성향
        - `wild`: 야생 구름 - 자유롭고 예측 불가능
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: 가입할 이메일 주소
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          description: 비밀번호 (최소 8자)
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: 로그인 이메일
          example: "user@example.com"
        password:
          type: string
          description: 비밀번호
    SocialLoginRequest:
      type: object
      required: [socialId, provider, email]
      properties:
        socialId:
          type: string
          description: 소셜 제공자에서 발급한 고유 사용자 ID
          example: "1234567890"
        provider:
          type: string
          enum: [GOOGLE, APPLE, KAKAO]
          description: 소셜 로그인 제공자
        email:
          type: string
          format: email
          description: 소셜 계정에 연결된 이메일
          example: "user@gmail.com"
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 사용자 고유 ID
        email:
          type: string
          description: 이메일 주소
        isPaid:
          type: boolean
          description: 유료 회원 여부
        createdAt:
          type: string
          format: date-time
          description: 가입 일시
    StepResponse:
      type: object
      description: |
        진단 테스트의 한 단계(Step).
        총 12단계로 구성되며, 각 단계에는 여러 질문이 포함됩니다.
        프론트엔드에서는 각 단계의 questions 중 랜덤 1개를 선택하여 표시합니다.
      properties:
        id:
          type: integer
          description: 단계 ID (1~12)
          example: 1
        title:
          type: string
          description: 단계 제목 (해당 단계의 주제)
          example: "감정 표현"
        emoji:
          type: string
          description: 단계를 대표하는 이모지
          example: "💭"
        questions:
          type: array
          description: 해당 단계에 속한 질문 목록 (이 중 랜덤 1개를 선택하여 사용)
          items:
            $ref: '#/components/schemas/QuestionResponse'
    QuestionResponse:
      type: object
      description: 개별 질문. 각 질문에는 4개의 선택지가 포함되며, 각 선택지는 특정 구름 유형과 매핑됩니다.
      properties:
        id:
          type: integer
          description: 질문 고유 ID
          example: 1
        text:
          type: string
          description: 질문 본문
          example: "기분이 안 좋을 때 나는..."
        options:
          type: array
          description: 선택지 목록 (보통 4개, 각각 다른 cloudType에 매핑)
          items:
            $ref: '#/components/schemas/OptionResponse'
    OptionResponse:
      type: object
      description: 질문의 선택지. 각 선택지를 선택하면 해당 cloudType에 1점이 부여됩니다.
      properties:
        text:
          type: string
          description: 선택지 텍스트
          example: "밝은 척 웃으며 넘긴다"
        cloudType:
          $ref: '#/components/schemas/CloudType'
    DiagnosisRequest:
      type: object
      required: [answers]
      description: |
        진단 분석 요청. 12개의 답변(각 단계에서 1개씩)을 배열로 전달합니다.
        각 답변에는 선택한 질문의 ID와 해당 선택지의 cloudType을 포함합니다.
      properties:
        answers:
          type: array
          description: 사용자 답변 배열 (12개)
          minItems: 12
          maxItems: 12
          items:
            $ref: '#/components/schemas/UserAnswer'
    UserAnswer:
      type: object
      required: [questionId, cloudType]
      description: 개별 답변. 어떤 질문에서 어떤 구름 유형을 선택했는지를 나타냅니다.
      properties:
        questionId:
          type: integer
          description: 응답한 질문의 ID
          example: 1
        cloudType:
          $ref: '#/components/schemas/CloudType'
    DiagnosisResult:
      type: object
      description: |
        진단 결과. 12개의 답변에서 가장 많이 선택된 구름 유형이 최종 결과로 반환됩니다.
        이 personaType 값으로 `/personas/profiles/{typeKey}` API를 호출하면 상세 프로필을 확인할 수 있습니다.
      properties:
        personaType:
          $ref: '#/components/schemas/CloudType'
    ChemistryResponse:
      type: object
      description: |
        두 구름 유형 간의 궁합 정보.
        두 유형이 만났을 때 어떤 하늘이 펼쳐지는지를 자연 현상 메타포로 표현합니다.
        content JSONB에서 추출한 상세 데이터(phenomenonName, vibeTags, storyBeats, premium)도 포함됩니다.
      properties:
        personaType1:
          $ref: '#/components/schemas/CloudType'
        personaType2:
          $ref: '#/components/schemas/CloudType'
        skyName:
          type: string
          description: 두 구름이 만들어내는 하늘의 영어 이름
          example: "Morning Light Through Fog"
        skyNameKo:
          type: string
          description: 두 구름이 만들어내는 하늘의 한국어 이름 (없을 수 있음)
          example: "아침 안개 속의 햇살"
        phenomenon:
          type: string
          description: 자연 현상 키워드 (glow, thunder, rain 등)
          example: "glow"
        narrative:
          type: string
          description: 두 유형의 관계에 대한 서사적 설명
          example: "햇살은 조용히 안개 곁에 머문다.\n안개는 서서히 마음의 결을 드러내고,\n햇살은 그 결을 지켜주고 싶어진다."
        warning:
          type: string
          description: 이 궁합에서 주의해야 할 점 (없을 수 있음)
          example: "빛이 서두르면, 안개는 숨을 곳을 찾는다."
        phenomenonName:
          type: object
          description: 현상 이름 (다국어)
          properties:
            en:
              type: string
              example: "Morning Light Through Fog"
            ko:
              type: string
              example: "아침 안개 속의 햇살"
        vibeTags:
          type: object
          description: 분위기 태그 (다국어 배열, 없을 수 있음)
          properties:
            en:
              type: array
              items:
                type: string
              example: ["tender", "patient"]
            ko:
              type: array
              items:
                type: string
              example: ["부드러운", "인내하는"]
        storyBeats:
          type: object
          description: 스토리 비트 (구조화된 서사 데이터, 없을 수 있음)
          additionalProperties: true
        premium:
          type: object
          description: 프리미엄 컨텐츠 (유료 전용 상세 서사, 없을 수 있음)
          additionalProperties: true
    PremiumCardResponse:
      type: object
      description: |
        프리미엄 카드 텍스트 템플릿.
        category별로 다양한 형태의 content를 가집니다.
        - 텍스트 배열: opening_patterns, softening_lines 등 (content: string[])
        - 마찰 패턴: friction_patterns (content: string)
        - 오해 패턴: misreading_patterns (content: [{signal, misread}])
        - 테마: themes (content: {bg, accent, sky, cloudOpacity})
      properties:
        id:
          type: integer
          description: 템플릿 고유 ID
          example: 1
        category:
          type: string
          description: 카테고리명
          example: "opening_patterns"
        subKey:
          type: string
          description: 하위 키 (friction/misreading의 패턴 타입, themes의 번호 등)
          example: "pace_mismatch"
        locale:
          type: string
          description: 언어 코드 (ko, en, 또는 _ = 언어 무관)
          example: "ko"
        content:
          description: 템플릿 내용 (카테고리에 따라 배열, 문자열, 객체 등 다양한 형태)
          oneOf:
            - type: array
              items:
                type: string
            - type: string
            - type: object
    TranslationAllResponse:
      type: object
      description: |
        locale의 전체 번역 데이터. namespace를 키로, 해당 namespace의 번역 맵을 값으로 가집니다.
        `{ namespace: { keyPath: value } }` 구조입니다.
      additionalProperties:
        type: object
        additionalProperties:
          type: string
      example:
        home:
          pageTitle: "Cloud Between Us — 우리 사이의 하늘"
          hero.headline: "사랑은, 날씨처럼 변해."
        test:
          back: "이전으로"
          next: "다음으로"
    TranslationNamespaceResponse:
      type: object
      description: |
        특정 namespace의 번역 데이터. keyPath를 키로, 번역 문자열을 값으로 가집니다.
        keyPath는 dot notation으로 중첩 구조를 표현합니다 (예: `hero.headline`, `unlockList.story`).
        프론트엔드에서 `.` 기준으로 split하여 원래 JSON 구조로 복원할 수 있습니다.
      additionalProperties:
        type: string
      example:
        pageTitle: "Cloud Between Us — 우리 사이의 하늘"
        hero.headline: "사랑은, 날씨처럼 변해."
        hero.cta: "우리 하늘 보기"
    PersonaProfileResponse:
      type: object
      description: |
        구름 유형의 상세 프로필.
        진단 결과 또는 유형 소개 화면에서 사용됩니다.
      properties:
        typeKey:
          $ref: '#/components/schemas/CloudType'
        name:
          type: string
          description: 구름 유형의 표시 이름
          example: "햇살 구름"
        emoji:
          type: string
          description: 대표 이모지
          example: "☀️"
        subtitle:
          type: string
          description: 한 줄 소개
          example: "세상을 밝게 비추는 따뜻한 빛"
        keywords:
          type: array
          description: 성격 키워드 (3~5개)
          items:
            type: string
          example: ["밝은", "따뜻한", "사교적인", "낙천적인"]
        lore:
          type: string
          description: 구름 유형에 대한 서사적 설명 (세계관 기반)
          example: "하늘 위에서 가장 먼저 빛을 받아들이는 구름..."
        strengths:
          type: array
          description: 이 유형의 강점 목록
          items:
            type: string
          example: ["사람들에게 에너지를 줌", "어디서든 분위기를 밝힘"]
        shadows:
          type: array
          description: 이 유형의 그림자(약점) 목록
          items:
            type: string
          example: ["자신의 감정을 숨기는 경향", "무리해서라도 밝게 행동함"]